+++ src/problems/P1.h
@@ -6,5 +6,5 @@
 #define A 3
 #define B 5
 
-int sum_ntuples(int, int);
+uint32_t sum_ntuples(uint16_t, uint16_t);
